# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Iris API
  description: API permettant de gérer les idées, votes et commentaires.
  version: 0.2.0
servers:
  - url: https://api.example.com
    description: Serveur de production (exemple)
  - url: http://localhost:8080
    description: Développement local
security:
  - bearerAuth: []
paths:
  /v1/images:
    post:
      summary: Importer une image
      description: "Importe une image .jpg ou .png (< 2Mo) et retourne l'URL publique pour la récupérer."
      tags: [Images]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '201':
          description: URL de l'image
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                required: [url]
        default:
          $ref: '#/components/responses/ErrorResponse'
  /images/{filename}:
    get:
      summary: Récupérer une image
      description: "Retourne le contenu de l'image."
      tags: [Images]
      security: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Image non trouvée
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: Erreur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnhancedError'
  schemas:
...
    EnhancedError:
      type: object
      properties:
        error_code:
          type: string
          description: Code d'erreur interne (ex. AUTH001, REQ000, SER000, ...)
        error_message:
          type: string
        error_http_code:
          type: integer
        details:
          description: Informations additionnelles liées à l'erreur
      required: [error_code]